+++ CHANGELOG.md
@@ -11,6 +11,10 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 - Handle dwebp re-coding errors
 
+### Fixed
+
+- Added workaround for wrong pre-downloads API format
+
 ## [3.10.0] - 17.06.2024
 
 ### Added